Add ORM entities for attributes and stat dependencies Fix bugs with movement

+++ src/main/kotlin/io/github/loggerworld/dto/response/character/PlayerClassResponse.kt
@@ -3,5 +3,5 @@ package io.github.loggerworld.dto.response.character
 data class PlayerClassResponse(
     var id: Byte = -1,
     var code: String = "",
-    var stats: Map<Byte, Double> = emptyMap(),
+    var stats: Map<Byte, Float> = emptyMap(),
 )
